requirements.txt
flask
flask-sqlalchemy
requests
psycopg2
